product_api/products/views.py
from django.shortcuts import render
from rest_framework import generics
from .models import Product
from .serializers import ProductSerializer

# Обработчик для списка и создания продуктов
class ProductListCreate(generics.ListCreateAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer

    def perform_create(self, serializer):
        serializer.save()

# Обработчик для отображения главной страницы
def index(request):
    return render(request, 'index.html')