requirements.txt
amqp==5.2.0
anyio==3.7.1
APScheduler==3.6.3
asgiref==3.7.2
atomicwrites==1.4.1
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.2.0
flake8==3.9.2
flake8-docstrings==1.6.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==4.13.0
iniconfig==2.0.0
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
logging==0.4.9.6
MarkupSafe==2.1.3
mccabe==0.6.1
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
Pillow==9.0.0
pluggy==0.13.1
prompt-toolkit==3.0.41
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.2.2
pyflakes==2.3.1
PyJWT==2.7.0
pyTelegramBotAPI==4.15.4
pytest==6.2.5
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.0
python-telegram-bot==20.3
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
style==1.1.0
telegram==0.0.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.12
webcolors==1.11.1
zipp==3.9.0
